HDiffPatcher와 AvatarRebuilder 연동 기능 추가

+++ Assets/Runtime/AvatarRebuilder/ReplaceModelAsset.cs
@@ -13,19 +13,23 @@ namespace VRSuya.Installer {
 
 	public class ReplaceModelAsset {
 
-		public void RequestReplaceModelAsset(string OldModelPath, string NewModelPath, int UndoGroupIndex) {
+		public void RequestReplaceModelAsset(string OldModelPath, string NewModelPath, bool Backuped, int UndoGroupIndex) {
 			string OldAvatarFullPath = Path.GetFullPath(OldModelPath);
 			string NewAvatarFullPath = Path.GetFullPath(NewModelPath);
 			string NewAvatarMetaFullPath = $"{NewAvatarFullPath}.meta";
-			string BackupAssetPath = GetBackupModelPath(OldModelPath);
-			string BackupFullPath = Path.GetFullPath(BackupAssetPath);
-			File.Copy(OldAvatarFullPath, BackupFullPath);
-			AssetDatabase.ImportAsset(BackupAssetPath);
-			CopyModelImporter CopyModelImporterInstance = new CopyModelImporter();
-			CopyModelImporterInstance.RequestCopyModelImporter(OldModelPath, BackupAssetPath, UndoGroupIndex);
+			if (!Backuped) {
+				string BackupAssetPath = GetBackupModelPath(OldModelPath);
+				string BackupFullPath = Path.GetFullPath(BackupAssetPath);
+				File.Copy(OldAvatarFullPath, BackupFullPath);
+				AssetDatabase.ImportAsset(BackupAssetPath);
+				CopyModelImporter CopyModelImporterInstance = new CopyModelImporter();
+				CopyModelImporterInstance.RequestCopyModelImporter(OldModelPath, BackupAssetPath, UndoGroupIndex);
+			}
 			File.Copy(NewAvatarFullPath, OldAvatarFullPath, overwrite: true);
-			File.Delete(NewAvatarFullPath);
-			File.Delete(NewAvatarMetaFullPath);
+			if (!Backuped) {
+				File.Delete(NewAvatarFullPath);
+				File.Delete(NewAvatarMetaFullPath);
+			}
 			AssetDatabase.Refresh();
 		}
